backend/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==11.1.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
python-dotenv==1.0.1
